requirements.txt
hindsight-client
hindsight-pydantic-ai
pydantic-ai[groq]
python-dotenv
rich
